+++ node/typings.d.ts
@@ -128,6 +128,12 @@ interface Address {
   checked?: boolean
 }
 
+interface B2BSetting {
+  autoApprove: boolean
+  defaultPaymentTerms: PaymentTerm[]
+  defaultPriceTables: [string]
+}
+
 interface UserArgs {
   id?: string
   roleId: string
@@ -143,4 +149,15 @@ interface UserArgs {
 interface PaymentTerm {
   name: string
   id: string
+}
+
+interface Price {
+  name: string
+  id: string
+}
+
+interface B2BSettingsInput {
+  autoApprove: boolean
+  defaultPaymentTerms: PaymentTerm[]
+  defaultPriceTables: Price[]
 }